feat: 온보딩 투타워 유저 타워 점수 체계 전환
- 유저 타워 입력과 저장 스키마를 0~1 실수 기준으로 통일
- base36 공유 코드를 새 점수 체계에 맞게 확장
- 예제 화면 슬라이더 UI와 테스트, 문서를 함께 갱신

+++ backend/services/onboarding-service/app/db/models.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime, timezone
 from typing import Any
 
-from sqlalchemy import JSON, DateTime, Index, Integer, SmallInteger, String, UniqueConstraint
+from sqlalchemy import JSON, DateTime, Float, Index, Integer, SmallInteger, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -28,15 +28,15 @@ class UserTowerProfileRecord(Base):
     user_id: Mapped[str] = mapped_column(String(64), nullable=False)
     profile_name: Mapped[str] = mapped_column(String(128), nullable=False)
     preferred_category_code: Mapped[str] = mapped_column(String(32), nullable=False)
-    budget_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    stability_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    subway_dependency_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    weekend_preference_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    evening_preference_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    resident_focus_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    worker_focus_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    rent_sensitivity_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
-    competition_tolerance_level: Mapped[int] = mapped_column(SmallInteger, nullable=False)
+    budget_level: Mapped[float] = mapped_column(Float, nullable=False)
+    stability_level: Mapped[float] = mapped_column(Float, nullable=False)
+    subway_dependency_level: Mapped[float] = mapped_column(Float, nullable=False)
+    weekend_preference_level: Mapped[float] = mapped_column(Float, nullable=False)
+    evening_preference_level: Mapped[float] = mapped_column(Float, nullable=False)
+    resident_focus_level: Mapped[float] = mapped_column(Float, nullable=False)
+    worker_focus_level: Mapped[float] = mapped_column(Float, nullable=False)
+    rent_sensitivity_level: Mapped[float] = mapped_column(Float, nullable=False)
+    competition_tolerance_level: Mapped[float] = mapped_column(Float, nullable=False)
     raw_answers: Mapped[dict[str, Any] | None] = mapped_column(JSON, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=_utc_now)
     updated_at: Mapped[datetime] = mapped_column(
